implemented working slider bars for each point type
