+++ unity game/SCS_vr/Assets/Prefabs/shorthair.prefab
@@ -674,7 +674,7 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1435174795211702}
-  m_LocalRotation: {x: 0, y: 0.7071068, z: 0, w: -0.7071068}
+  m_LocalRotation: {x: 0, y: -0.7071068, z: 0, w: -0.7071068}
   m_LocalPosition: {x: 2.5230618, y: -0.562, z: -2.6962404}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
