+++ unity game/SCS_vr/Assets/Scripts/SliderScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a1abe5783b9e51c49a2706d4ed387d70
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
