+++ unity game/SCS_vr/Assets/Scripts/SliderScripts/GreenSlider.cs
@@ -3,24 +3,27 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class Sliders : MonoBehaviour {
+public class GreenSlider : MonoBehaviour
+{
 
     public Slider slider;
     public Text valueText;
     private float value;
 
-	// Use this for initialization
-	void Start () {
+    // Use this for initialization
+    void Start()
+    {
         PointsController pointsScript = GameObject.Find("PointsController").GetComponent<PointsController>();
-        value = pointsScript.pointsData.money;
-        valueText.text = "$" + value.ToString();
+        value = pointsScript.pointsData.green;
+        valueText.text = value.ToString();
         slider.value = value;
-	}
-	
-	void Update () {
+    }
+
+    void Update()
+    {
         PointsController pointsScript = GameObject.Find("PointsController").GetComponent<PointsController>();
-        value = pointsScript.pointsData.money;
-        valueText.text = "$" + value.ToString();
-        slider.value = value/100;
+        value = pointsScript.pointsData.green;
+        valueText.text = value.ToString();
+        slider.value = value / 100;
     }
 }
