+++ unity game/SCS_vr/Assets/Scripts/SliderScripts/GreenSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e8824ffd82a487947a5bb14292bdea40
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
