+++ unity game/SCS_vr/Assets/Scripts/SliderScripts/PopSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be2996270026d2a4cbe7f12294154b25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
